Fixed table sorting when array is passed to dataIndex
