+++ src/components/Table/components/HeaderCell/HeaderCellMenu.jsx
@@ -2,6 +2,7 @@ import React, { useRef } from "react";
 
 import { isPresent } from "neetocist";
 import { Check, MenuHorizontal } from "neetoicons";
+import { equals } from "ramda";
 import { useTranslation } from "react-i18next";
 
 import Dropdown from "components/Dropdown";
@@ -75,7 +76,7 @@ const HeaderCellMenu = ({
               >
                 <span>{getLocale(i18n, t, "neetoui.table.ascending")}</span>
                 {sortedInfo.order === TABLE_SORT_ORDERS.asc &&
-                  sortedInfo.field === column.dataIndex && (
+                  equals(sortedInfo.field, column.dataIndex) && (
                     <Check className="neeto-ui-text-success-500" size={20} />
                   )}
               </MenuItem.Button>
@@ -93,7 +94,7 @@ const HeaderCellMenu = ({
               >
                 <span>{getLocale(i18n, t, "neetoui.table.descending")}</span>
                 {sortedInfo.order === TABLE_SORT_ORDERS.desc &&
-                  sortedInfo.field === column.dataIndex && (
+                  equals(sortedInfo.field, column.dataIndex) && (
                     <Check className="neeto-ui-text-success-500" size={20} />
                   )}
               </MenuItem.Button>
